requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.15.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
coreapi==2.3.3
coreschema==0.0.4
croniter==2.0.5
cryptography==43.0.0
defusedxml==0.8.0rc2
Deprecated==1.2.14
distro==1.9.0
dj-rest-auth==2.2.4
Django==4.2
django-allauth==0.50.0
django-appconf==1.0.6
django-autoslug==1.9.8
django-ckeditor==5.9.0
django-constance==2.9.0
django-cors-headers==4.4.0
django-debug-toolbar==3.4.0
django-environ==0.11.2
django-extensions==3.1.5
django-filter==24.3
django-intl-tel-input==0.3.1
django-js-asset==2.2.0
django-menu-generator==1.0.4
django-model-utils==4.1.1
django-modeltranslation==0.18.2
django-pdfkit==0.3.1
django-phonenumber-field==4.0.0
django-phonenumbers==1.0.1
django-picklefield==3.2
django-query-builder==2.0.1
django-ratelimit==4.1.0
django-redis==5.2.0
django-rest-framework==0.1.0
django-rosetta==0.9.8
django-rq==2.5.1
django-select2==7.2.3
django-simple-history==3.7.0
django-storages==1.14.4
django-tables2==2.3.1
django-taggit==1.2.0
django-templated-mail==1.1.1
django-userforeignkey==0.3.0
django-uuidfield==0.5.0
django-versatileimagefield==3.1
django-widget-tweaks==1.4.12
django_reverse_admin==2.9.6
djangorestframework==3.15.2
djangorestframework-bulk==0.2.1
djangorestframework-datatables==0.7.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.1
drf-nested-forms==1.1.7
drf-writable-nested==0.6.3
environ==1.0
flatbuffers==24.3.25
fleming==0.7.0
google-api-core==2.20.0
google-api-python-client==2.154.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
humanfriendly==10.0
idna==3.7
imageio==2.35.1
itypes==1.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.0
oauthlib==3.2.2
onedrivesdk==2.0.1
onnxruntime==1.19.2
openai==1.46.0
opencv-python-headless==4.10.0.84
packaging==24.1
pdfkit==0.6.0
phonenumbers==8.13.39
pillow==10.4.0
platformdirs==4.3.6
polib==1.2.0
pooch==1.8.2
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMatting==1.1.12
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-decouple==3.6
python-magic==0.4.27
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2024.1
redis==4.3.3
referencing==0.35.1
rembg==2.0.59
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rq==1.16.2
rq-scheduler==0.11.0
rsa==4.9
scikit-image==0.24.0
scipy==1.14.1
setuptools==73.0.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sorl-thumbnail==12.8.0
sorl-thumbnail-serializer-field==0.2.1
sqlparse==0.5.1
sympy==1.13.3
tifffile==2024.9.20
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.2.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.17.6
wheel==0.43.0
whitenoise==6.7.0
wrapt==1.16.0
